rename playwright tests: d2d to e2e, lib-ui to ladle
