+++ frontend/README.md
@@ -52,14 +52,14 @@ Browser tests using Playwright:
 
 ```sh
 # tests for ui components using ladle:
-npm run e2e:lib-ui
+npm run test:ladle
 
 # tests for frontend with backend and database
 # tests use database at backend/target/debug/playwright.sqlite
 # build frontend, build backend, setup fresh seeded database:
-npm run e2e:d2d
+npm run test:e2e
 # run tests, expect builds and database to be available:
-npm run e2e:d2d-dev
+npm run test:e2e-dev
 ```
 
 ### UI Component development
