+++ frontend/e2e-tests/polling-station-crud.e2e.ts
@@ -1,9 +1,9 @@
 import { expect } from "@playwright/test";
 import { PollingStationFormPgObj } from "e2e-tests/page-objects/polling_station/PollingStationFormPgObj";
 
-import { PollingStationListEmptyPgObj } from "../page-objects/polling_station/PollingStationListEmptyPgObj";
-import { PollingStationListPgObj } from "../page-objects/polling_station/PollingStationListPgObj";
 import { test } from "./fixtures";
+import { PollingStationListEmptyPgObj } from "./page-objects/polling_station/PollingStationListEmptyPgObj";
+import { PollingStationListPgObj } from "./page-objects/polling_station/PollingStationListPgObj";
 
 test.describe("Polling station CRUD", () => {
   test("it redirects correctly after successful create of first polling station of an election", async ({
