+++ frontend/package.json
@@ -15,9 +15,9 @@
     "prettier": "prettier --ignore-unknown --write .",
     "preview": "vite preview",
     "test": "vitest",
-    "e2e:lib-ui": "playwright test -c playwright.lib-ui.config.ts",
-    "e2e:d2d": "cross-env LOCAL_CI=true playwright test -c playwright.d2d.config.ts",
-    "e2e:d2d-dev": "playwright test -c playwright.d2d.config.ts",
+    "test:e2e": "cross-env LOCAL_CI=true playwright test -c playwright.e2e.config.ts",
+    "test:e2e-dev": "playwright test -c playwright.e2e.config.ts",
+    "test:ladle": "playwright test -c playwright.ladle.config.ts",
     "ladle": "ladle serve",
     "build:ladle": "ladle build",
     "start:msw": "cross-env API_MODE=mock vite",
